requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.10.1
click==8.1.7
Django==4.2.6
django-admin==2.0.2
django-debug-toolbar==4.2.0
django-excel-response2==3.0.6
django-rest-framework==0.1.0
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
excel-base==1.0.4
inflection==0.5.1
isoweek==1.3.3
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rest-framework-simplejwt==0.0.2
rpds-py==0.10.6
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
tzlocal==5.2
uritemplate==4.1.1
xlwt==1.3.0
